Moved shared files to components
