+++ src/lib/components/single-line-chart.svelte
