+++ svelte.config.js
@@ -17,7 +17,6 @@ const config = {
 				alias: {
 					$src: path.resolve('./src'),
 					$lib: path.resolve('./src/lib'),
-					$shared: path.resolve('./src/lib/shared'),
 					$components: path.resolve('./src/lib/components'),
 					$stores: path.resolve('./src/lib/stores')
 				}
